+++ api/api.py
@@ -11,7 +11,6 @@ from rest_framework.viewsets import ReadOnlyModelViewSet
 from apps.budget.choices import ExpenseGroupChoices, RevenueGroupChoices
 from apps.budget.models import Budget, TransparencyIndex, Expense, Revenue, BudgetSummary, Category, CategoryMap
 from apps.geo.models import Country
-from common.methods import apply_change, apply_change_euro
 
 
 class CountrySerializer(serializers.ModelSerializer):
@@ -423,34 +422,6 @@ class BudgetViewset(ReadOnlyModelViewSet):
                 agregateExpense['percent'] = (expense['budget_aggregated'] / totalbudgetSummaryCountry) * 100;
                 agregateExpenses.append(agregateExpense)
         return Response(agregateExpenses)
-
-
-    @action(detail=False)
-    def summary_budget_per_year_all_country(self, request, pk=None):
-        year = request.query_params.get('year', 2018)
-        base_currency = request.query_params.get('base_currency')
-        aggregate_budgets_summery = []
-        
-        #UpGreat code to new 
-        #Identify an Alteraction to pull
-
-        budget_summerys = BudgetSummary.objects.filter(budget__year=year)
-
-        """ Puchar o as Rates """
-        rate = apply_change(base_currency)
-
-        for budgetSummery in budget_summerys:
-            tupla = {}
-            
-            tupla['country'] = budgetSummery.budget.country.name
-            tupla['budget'] = budgetSummery.expense_organic_budget if budgetSummery.expense_functional_budget is None else budgetSummery.expense_functional_budget
-            tupla['budget_euro'] = budgetSummery.expense_organic_budget / rate.get(budgetSummery.budget.currency) if budgetSummery.expense_functional_budget is None else budgetSummery.expense_functional_budget / rate.get(budgetSummery.budget.currency)
-            tupla['year'] = year
-            tupla['source'] = 'Organic' if budgetSummery.expense_functional_budget is None else 'Functional'
-
-            aggregate_budgets_summery.append(tupla)
-
-        return Response(aggregate_budgets_summery)
 
     @action(detail=False)
     def expense_by_country(self, request, pk=None):
