+++ api/urls.py
@@ -4,11 +4,15 @@ from rest_framework.routers import DefaultRouter
 
 from api.api import BudgetViewset
 from api.api_admin import AdminViewset
+from api.students.angola_estima import AngolaEstimaViewset
 
 router = DefaultRouter()
 
 router.register(r'api-admin', AdminViewset, basename='api_admin')
 router.register(r'budgets', BudgetViewset, basename='budgets')
+
+# Students
+router.register(r'angola-estima', AngolaEstimaViewset, basename='angola_estima_api')
 
 urlpatterns = [
     url(r'^', include(router.urls)),
